refactor: Move tool injectables into seperate file
